maint: Remove outdated test backup file
The `data-table.testbak.tsx` file, which appears to be a backup or outdated version of the `data-table` component tests, has been deleted. This file was not in use and did not affect the current test suite or functionality of the `data-table` component. The deletion helps maintain a cleaner codebase and reduces potential confusion among developers.